requirements.txt
aioredis==2.0.1
anyio==3.6.1
astroid==2.12.5
async-timeout==4.0.2
attrs==22.1.0
black==22.6.0
cleo==1.0.0a4
click==8.1.3
crashtest==0.3.1
Deprecated==1.2.13
dill==0.3.5.1
fastapi==0.81.0
h11==0.13.0
hiredis==2.0.0
httptools==0.4.0
idna==3.3
iniconfig==1.1.1
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.7.0
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
platformdirs==2.5.2
pluggy==1.0.0
pptree==3.1
py==1.11.0
pydantic==1.9.2
pylev==1.4.0
pylint==2.15.0
pyparsing==3.0.9
pytest==7.1.2
python-dotenv==0.20.0
python-ulid==1.1.0
PyYAML==6.0
redis==4.3.4
redis-om==0.0.27
six==1.16.0
sniffio==1.2.0
starlette==0.19.1
tomli==2.0.1
tomlkit==0.11.4
types-redis==4.3.19
types-six==1.16.19
typing_extensions==4.3.0
uvicorn==0.18.3
uvloop==0.16.0
watchfiles==0.16.1
websockets==10.3
wrapt==1.14.1
